/go/libraries/doltcore/{doltdb,sqle}: create dolt_ci tables on dolt init
