+++ go/libraries/doltcore/sqle/database.go
@@ -413,28 +413,6 @@ func (db Database) getTableInsensitive(ctx *sql.Context, head *doltdb.Commit, ds
 	var dt sql.Table
 	found := false
 	switch lwrName {
-	//case doltdb.WorkflowsTableName:
-	//	backingTable, _, err := db.getTable(ctx, root, doltdb.WorkflowsTableName)
-	//	if err != nil {
-	//		return nil, false, err
-	//	}
-	//	if backingTable == nil {
-	//		dt, found = NewEmptyWorkflowsTable(ctx), true
-	//	} else {
-	//		versionableTable := backingTable.(dtables.VersionableTable)
-	//		dt, found = NewWorkflowsTable(ctx, db.ddb, versionableTable), true
-	//	}
-	//case doltdb.WorkflowEventsTableName:
-	//	backingTable, _, err := db.getTable(ctx, root, doltdb.WorkflowEventsTableName)
-	//	if err != nil {
-	//		return nil, false, err
-	//	}
-	//	if backingTable == nil {
-	//		dt, found = NewEmptyWorkflowEventsTable(ctx, db), true
-	//	} else {
-	//		versionableTable := backingTable.(dtables.VersionableTable)
-	//		dt, found = NewWorkflowEventsTable(ctx, db, versionableTable), true
-	//	}
 	case doltdb.LogTableName:
 		if head == nil {
 			var err error
